Add support for reading the default model from a config file
This commit adds support for reading the default model from a config file
located at `~/.aichat/config.yml`. If the `model` flag is not specified when
running the program, it will try to read the default model from the config file.
If the config file does not exist or does not contain a `model` field, it will
use the default model `gogpt.GPT3Dot5Turbo`.